requirements.txt
alembic==1.10.2
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-Login==0.6.2
Flask-Migrate==4.0.4
flask-paginate==2022.1.8
Flask-SQLAlchemy==3.0.3
Flask-Uploads==0.2.1
Flask-WTF==1.1.1
# greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
# mysqlclient==2.1.1
python-slugify==8.0.1
SQLAlchemy==2.0.5.post1
text-unidecode==1.3
typing_extensions==4.5.0
Werkzeug==2.2.3
WTForms==3.0.1
